refactor(heureka): remove overlooked index.ts
