+++ apps/heureka/src/components/Services/ServicesList/index.tsx
@@ -10,7 +10,7 @@ import { MessagesProvider } from "@cloudoperators/juno-messages-provider"
 import { ServiceListItem } from "./ServiceListItem"
 import { EmptyDataGridRow } from "../../common/EmptyDataGridRow"
 import { ServicePanel } from "./ServicePanel"
-import { ServiceType } from "../Services"
+import { ServiceType } from ".."
 import { useFetchServices } from "../useFetchServices"
 import { FilterSettings } from "../../common/Filters/types"
 import { useDispatch } from "../../../store/StoreProvider"
