+++ apps/heureka/src/components/Services/utils.ts
@@ -14,7 +14,7 @@ import {
   ServiceFilter,
   GetServiceImageVersionsQuery,
 } from "../../generated/graphql"
-import { ServiceType } from "./Services"
+import { ServiceType } from "./index"
 import { FilterSettings, ServiceFilterReduced } from "../common/Filters/types"
 import { IssuesCountsType } from "./useFetchServicesIssuesCounts"
