+++ apps/heureka/src/components/common/ServiceImageVersions/index.tsx
@@ -18,7 +18,7 @@ import { EmptyDataGridRow } from "../EmptyDataGridRow"
 import { useFetchServiceImageVersions } from "../../Services/useFetchServiceImageVersions"
 import { useDispatch } from "../../../store/StoreProvider"
 import { ActionType, UserView } from "../../../store/StoreProvider/types"
-import { ServiceType } from "../../Services/Services"
+import { ServiceType } from "../../Services"
 import ServiceImageVersionsItem from "./ServiceImageVersionsItem"
 import { ServiceImageVersion } from "../../Services/utils"
 import SectionContentHeading from "../SectionContentHeading"
